Attempted to visualize the predictions, and yeah, the values generated by the RatingsModel simply do not match real world values, even after training for hours.
My guess, is I will need to first calculate the margins of error, then calculate the renewal odds. Then I will need a method which experimentally try to find how much the ratings must be increased or decreased in order to be the lowest rating value that could achieve a renewal.

The problem is needing to do this for both ratings and viewers both, as well as finding a number that is representative of the entire season, since that was the entire purpose of the RatingsModel, to find a value that best represents the season as a whole. This will be a difficult problem to solve.