Decided to at least temporarliy disable "Find Similar Show" and "Prediction Breakdown" tabs, as Prediction Breakdown is difficult to implement and not often used. Find Similar Shows can be interesting to estimate the renewal threshold based on where similar shows were renewed and cancelled, but was rarely used. I may eventually add them back, but for now I don't feel like implementing them.
The final tab, Modify Factors WILL need to be implemented at some time, but because it will result in resetting the network's evolution model, I have decided to wait to implement that until I actually need it. This will allow you to rename, manually sort, delete, and add new factors for the entire network. Adding is the most complicated as it requires making a modification for that factor for every single show.

At this point, the app is functionally complete for what I will need of it until factor modifications are needed. At least, until I can think of any new changes that might be necessary down the line.