Modified EvolutionController to reset Accuracy and Error for all models in all networks before beginning the first generation. I had previously assumed that by the end of the last generation, these should have been null as new models are created from their parents, but for some reason, this was not the case.
Actually, I just checked and the copy constructor for PredictionModel was indeed copying Accuracy and Error. It turns out these values are only made null if a mutation occurs, so sometimes the model can be left with the same accuracy and error as its parent, but this is only a problem when starting a new EvolutionController, as show data could have changed, making these values inaccurate.